Updated the bot (NOT TESTED)

+++ modules/nuke/cog.py
@@ -1,10 +1,8 @@
+
 from discord.ext import commands
 import discord
-import discord_slash
-from discord_slash import SlashCommand, SlashContext
-from discord_slash.utils.manage_commands import create_choice, create_option
 from discord.ext.tasks import loop
-from discord_slash import cog_ext
+
 import datetime
 
 
@@ -15,30 +13,37 @@ class nuke(commands.Cog):
     @commands.command(name='nuke')
     @commands.guild_only()
     @commands.has_permissions(manage_messages=True)
-    async def Nuke(self, ctx: SlashContext, channel: discord.TextChannel = None):
-        try:
-            if channel == None:
-                channel = ctx.channel
-            nuke_channel = discord.utils.get(
-                ctx.guild.channels, name=channel.name)
+    async def Nuke(self, interaction: discord.Interaction, channel: discord.TextChannel = None):
+        
+        bot_member = interaction.guild.get_member(self.bot.user.id)
+        if not bot_member.guild_permissions.manage_channels:
+            return await interaction.response.send_message("I do not have the permission to manage channels.", ephemeral=True)
+            
+        
+        if not interaction.user.guild_permissions.manage_channels:
+            embed = discord.Embed(
+                title="Permission Error",
+                description=f"{interaction.user.mention}, you don't have enough permissions to use this command.",
+                color=discord.Color.red()
+            ).set_footer(
+                text=f"Requested by {interaction.user.name}",
+                icon_url=interaction.user.avatar
+            )
+            return await interaction.response.send_message(embed=embed, ephemeral=True)
+        if channel == None:
+            channel = interaction.channel
+        nuke_channel = discord.utils.get(
+            interaction.guild.channels, name=channel.name)
 
-            if nuke_channel is not None:
-                new_channel = await nuke_channel.clone(reason="Has been Nuked!")
-                await nuke_channel.delete()
-                await new_channel.send("THIS CHANNEL HAS BEEN NUKED!")
-                await new_channel.send("https://tenor.com/view/xqc-spotify-mood-want-a-break-bruh-gif-24599734")
+        if nuke_channel is not None:
+            new_channel = await nuke_channel.clone(reason="Has been Nuked!")
+            await nuke_channel.delete()
+            await new_channel.send("THIS CHANNEL HAS BEEN NUKED!")
+            await new_channel.send("https://tenor.com/view/xqc-spotify-mood-want-a-break-bruh-gif-24599734")
 
-            else:
-                await ctx.send(f"No channel named {channel.name} was found!")
-        except Exception as e:
-            embed = discord.Embed(title=":x: Command Error",
-            colour=0x992D22)  # Dark Red
-            embed.add_field(name="Error", value=e)
-            embed.add_field(name="Guild", value=ctx.guild)
-            embed.add_field(name="Channel", value=ctx.channel)
-            embed.add_field(name="User", value=ctx.author)
-            embed.timestamp = datetime.datetime.utcnow()
-            await ctx.reply(embed=embed)
+        else:
+            await interaction.response.send_message(f"No channel named {channel.name} was found!", ephemeral=True)
+
 
 
 def setup(bot):
